requirements.txt
aiogram==3.x
SQLAlchemy
psycopg2-binary
python-dotenv